"Fixed" PlaceTowerOnGrid
The original issue was that the tower was never being built before checking path, so it thought it could make the path before placing the tower.
The current issue is the same thing, but the problem is that unity doesnt actually place the instantiated object until after the frame is finished. Meaning we either have to figure out a way to get it to place before the frame is over (I tried and seems like nobody knows how) to get the cut out to happen, or use the grid array since that is updated realtime.